==> notebooks/06-data-preparation.ipynb <==
# Data preparation:

This notebook encompasses various essential steps to analyze the stock market data comprehensively. Initially, I gather and process the stock market data. Next, I combine this data with the data acquired in previous notebooks from FRED, topic modeling, and sentiment analysis. Additionally, I integrate another valuable data source, capturing economic policy uncertainty as expressed in news and media.

The complete process will be thoroughly explained within the notebook, providing in-depth insights into each stage of the analysis.

import numpy as np
import pandas as pd
import yfinance as yf

import warnings

pd.set_option('display.max_columns', 100)

# Set the start and end date
start_date = '1993-01-01'
end_date = '2023-07-27'
 
# Set the ticker
sp = '^GSPC'
dji = '^DJI'
ru2000 = '^RUT' #Russel 2000
ru3000 = '^RUA' # Russel 3000 --> captures 98% of the stock market
 
# Get the data
def stock_data(ticker, start=start_date, end=end_date):
    return yf.download(ticker, start, end)

sp = stock_data(sp)
dji = stock_data(dji)
ru2000 = stock_data(ru2000)
ru3000 = stock_data(ru3000)


the `sp` dataset provides information about the S&P 500 index's price movements and trading activity from 1993 to 2023. The columns contain data related to opening price, highest price, lowest price, closing price, adjusted closing price, and trading volume.


sp.head()

# Removing columns that are not needed in the analysis
sp.drop(columns=['High', 'Low', 'Close', 'Open'], inplace=True)

# Calculate the percentage change in the adjusted closing price and store it in a new column 'pct_change_adj_close'.
# 'sp['Adj Close']' represents the column containing the adjusted closing prices of the financial instrument.
sp['pct_change_adj_close'] = sp['Adj Close'].pct_change() * 100

# Calculate the percentage change in the trading volume and store it in a new column 'pct_change_volume'.

sp['pct_change_volume'] = sp['Volume'].pct_change() * 100
sp.head()

# Create a new column 'lag_pct_change_adj_close' and shift the values of 'pct_change_adj_close' one step forward.
# The 'shift()' function moves the values one position backward, effectively creating a lagged version of the column.
# The 'lag_pct_change_adj_close' column will now contain the percentage change values of the adjusted closing price
# for the previous trading day.
sp['lag_pct_change_adj_close'] = sp['pct_change_adj_close'].shift()

# Create a new column 'lag_pct_change_volume' and shift the values of 'pct_change_volume' one step forward.
sp['lag_pct_change_volume'] = sp['pct_change_volume'].shift()

sp.head()

# 'lag_ma15_pct_change_adj_close' calculates the rolling 15-day moving average of 'lag_pct_change_adj_close'.
sp['lag_ma15_pct_change_adj_close'] = sp['lag_pct_change_adj_close'].rolling(15).mean()
sp.head(10) 

Next, I import the FOMC minutes data, which contains the topic modeling data and the sentiment analyses data.

fomc_topic = pd.read_pickle('../data/fomc_topic_modeling.pkl')

fomc_topic.drop(columns=['minutes_paragraphs', 'paragraphs_length', 'minutes_text',
       'text_length', 'topic_weight', 'topic_score', 'doc2bow'], inplace=True)

fomc_topic.head()

fomc_laughran_mcdonald = pd.read_pickle('../data/fomc_raw_laughran_mcdonald_sentiment.pkl')
fomc_laughran_mcdonald = fomc_laughran_mcdonald[['sentiment_mc']]
fomc_laughran_mcdonald.head()

fomc_zstc_sentiment = pd.read_pickle('../data/fomc_raw_zstc_sentiment.pkl')
fomc_zstc_sentiment.drop(columns='fomc_text', inplace=True)
fomc_zstc_sentiment.head()

In the next step, I combine the three dataframes, merging them into a single comprehensive dataset. This merged dataset will contain all the information gathered from the various FOMC meetings.

fomc = fomc_topic.merge(fomc_laughran_mcdonald, how = 'inner', left_index=True, right_index=True) \
                 .merge(fomc_zstc_sentiment, how = 'inner', left_index=True, right_index=True)

Since the stock market operates only on weekdays, it's essential to check if all FOMC meetings happened during weekdays too. To do this, I add a new column called "dayofweek" to indicate the meeting day.

After reviewing the data, I find that one of the meetings took place on a Sunday, which could cause us to miss valuable information. To avoid this, I move the data from that Sunday meeting to the following Monday. By doing so, we ensure that we don't lose any important data, and all meetings align with the market's weekday schedule.

# The 'strftime('%A')' function is used to convert the date from the index to the full weekday name.
# For example, if the index contains '2023-07-30', the 'dayofweek' column for that row will have the value 'Saturday'.
fomc['dayofweek'] = fomc.index.strftime('%A')

fomc['dayofweek'].value_counts()


fomc.reset_index(names='date', inplace=True)

sunday_row = fomc['dayofweek'] =='Sunday'
fomc[sunday_row]

# For those meetings on Sundays, add one day to their 'date' values to move them to the following Monday.
# The 'pd.Timedelta(1, "d")' represents a time duration of one day, and it is added to the 'date' column.
fomc.loc[sunday_row, 'date'] = fomc.loc[sunday_row, 'date'] + pd.Timedelta(1, "d")

fomc.loc[sunday_row, 'date']

fomc.loc[sunday_row, 'dayofweek'] = 'Monday'

fomc['dayofweek'].value_counts()

fomc.set_index('date', inplace=True)

Next, I merge sp data with fomc and drop `dayofweek` column.

# Perform a left merge between the DataFrames 'sp' and 'fomc_topic' based on their index values.
data = pd.merge(sp, fomc, how='left', left_index=True, right_index=True)

# Drop the 'dayofweek' column from the merged DataFrame 'data'.
data = data.drop(columns='dayofweek')

data.head()

print(sp.shape)
print(fomc_topic.shape)

data.notna().sum()

> A possible concern is whether FOMC meetings might fall on weekdays that are holidays, which is unlikely but still important to consider. However, upon reviewing the data, we find that all 244 meeting dates are included in the final dataset.

Next, I add the FRED data to our existing dataset.

fred = pd.read_pickle('../data/fred.pkl')
fred.head(3)

FRED data contains various indicators with different frequency levels. While some indicators have daily data, others might have quarterly or monthly data, which can fall on weekends. To ensure that no data points are missed, I apply two methods from the pandas package: `resample` and `ffill`.

Using the `resample` method, I create a complete sample that covers all days from 1993 to 2023. Then, the `ffill` method is used to forward-fill any missing data points, ensuring that all relevant information is included in the final dataset. 

fred.describe()

(fred['10_year_treasury_yeild']==0).sum()

Fortunately, there are no values equal to zero in the FRED dataset. Therefore, we can create a complete sample spanning from 1993 to 2023 by utilizing Pandas `resample` method. This function will help us generate a dataset containing all the days within this timeframe. After creating the sample, we will replace any zero values with `NaN`  again, to maintain consistency with the rest of the data.

fred = fred.resample('D').sum()
fred.head(3)

fred.replace(0, np.nan, inplace=True)
fred.head(3)

Finally, I apply the `ffill` method for forward filling. This process ensures that any gaps in the data are filled with the most recent available value, completing our dataset and preparing it for further analysis.

fred.ffill(inplace=True)

fred.head(3)

At last, it is time to introduce and include the final indicator to the dataset, Economic Policy Uncertainty. The Economic Policy Uncertainty (EPU) Index is an economic indicator that measures the level of uncertainty in a country's economy arising from policy-related factors. It is designed to capture the extent to which economic policymakers (e.g., governments, central banks) are creating uncertainty through their decisions or the changing policy environment. The index helps in quantifying and monitoring economic uncertainty, which can have significant implications for businesses, investors, and policymakers.

The EPU Index composed of three types of underlying components. The first component quantifies newspaper coverage of policy-related economic uncertainty from ten large newspapers in the United States. The second component involves the number of federal tax code provisions set to expire, and the third component utilizes the dispersion between individual forecasters' predictions about future macroeconomic variables from the Federal Reserve Bank of Philadelphia's Survey of Professional Forecasters. These components collectively form the policy-related uncertainty index. ([source](http://www.policyuncertainty.com/index.html))

epu = pd.read_excel('../data/us_policy_uncertainty.xlsx', sheet_name='Main Index')
epu = epu.iloc[0:-1, :]
epu.head()

epu.dtypes

# Combine the 'Year' and 'Month' columns to create a new 'date' column in the DataFrame.
# The 'date' column will represent the first day of each month in the format 'YYYY-MM-1'.
epu['date'] = epu['Year'] + '-' + epu['Month'].astype(int).astype(str) + '-1'

# Convert the 'date' column to the datetime data type using the 'pd.to_datetime()' function.
epu['date'] = pd.to_datetime(epu['date'])

# Set the 'date' column as the new index of the DataFrame 'us_policy_uncertainty'.
epu.set_index('date', inplace=True)

# Drop the 'Year' and 'Month' columns from the DataFrame 'us_policy_uncertainty'.
epu.drop(columns=['Year', 'Month'], inplace=True)
epu.head(3)

epu.rename(columns={'Three_Component_Index':'epu_all',
                    'News_Based_Policy_Uncert_Index':'epu_news'}, inplace=True)

EPU, being recorded monthly, requires a similar cleaning approach as I used for FRED data. To achieve this, I will utilize the `resample` and `ffill` methods once again. By resampling the data, I create a complete daily sample that covers the entire duration. Then, the "ffill" method will be applied to fill in missing values. This process allows us to effectively handle any gaps or irregularities in the EPU data.

epu = epu.resample('D').sum()
epu.head(3)

epu.replace(0, np.nan, inplace=True)
epu.head(3)

epu.ffill(inplace=True)
epu.head(3)

> After importing and processing the economic policy uncertainty data from its original source, I made an interesting observation. It turns out that the index is also available in FRED, and it can be accessed using the series ID USEPUINDXD. 

Finally, I merge the data from `FRED` and the economic policy uncertainty (`EPU`) with the main dataframe, which is called `data`.  Additionally, I apply the "ffill" method once more to fill in missing values in the FOMC variables within the final sample. After that, all remaining missing values are removed.

data = pd.merge(data, fred, how='left', left_index=True, right_index=True) \
         .merge(epu, how='left', left_index=True, right_index=True)
data.ffill(inplace=True)
data.dropna(inplace=True)

data.head(3) 

The last set of data preparation steps are to create a dummy variable that indicate whether the stock prices has increases or not in each trading day. this variable will serve as the target variable in the analysis. Then, I remove, Volume and pct_change_volume from the dataset to avoid inputing forward looking data to the model.

The final steps of data preparation involve creating a dummy variable to indicate whether the stock prices have increased or not on each trading day. This variable will serve as the target variable for the analysis. Additionally, to avoid inputting forward-looking data into the model, I remove `pct_change_volume` from the dataset. By doing so, I ensure that the model uses only historical information for analysis and predictions, maintaining the integrity of the analysis and avoiding any data leakage issues. 

data['price_increase'] = np.where(data['pct_change_adj_close']>=0, 1, 0)

data.drop(columns=['pct_change_volume'], inplace=True)

data.head(3)

data.to_pickle('../data/full_data.pkl')
